server/requirements.txt
Django
djangorestframework
Pillow